pyproject.toml
[tool.poetry]
name = "es-client"
version = "0.1.0"
description = "gRPC client for EventStore DB"
authors = ["Andrii Kohut <kogut.andriy@gmail.com>"]
license = "MIT"

packages = [
    { include = "esdb" },
]

exclude = ["EventStore", "certs", "es_config", "protos"]

[tool.poetry.dependencies]
python = "3.10"

[tool.poetry.dev-dependencies]
grpcio-tools = "^1.47.0"
mypy = "^0.971"
mypy-protobuf = "^3.2.0"
black = "^22.6.0"
pytest = "^7.1.2"
isort = "^5.10.1"
pytest-asyncio = "^0.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = "esdb/generated"

[tool.isort]
profile = "black"
skip_glob = "esdb/generated/*"
